Module 4B Update
Implemented Functioning Login, Password Reset, and Create Account System for the webpage.

Webpage needs a little work around Login works on index.html but is not implemented for account.html; also need to decide if guests can use the website or not (as it will affect functionality)

Email Confirmation System needs implemented still and for now password reseting works but is only good for testing until further security is implemented. The same is true with creating accounts as emails won't be confirmed to be owned by anybody, but passwords themselves are secured in the database with hashing.